Started Complete Reorganization of Code
In the interest of making this script more open and generic for others to be able to utilize, I've started reorganizing and redoing the entire code so that users are presented with more options. In theory i would like this script to be able to be utilized on more than just the linux distro that i am using.