Added Loop and Remembering Entered Parameters
Added a loop so that when a session is done it will bring you back to the screen to be able to start another session. This is mainly beneficial for when invalid information gets entered. Along with that it will now remember, in the same session, the Server, Port, Gateway and Username. So, those will not have to be re-entered.