Added MS Teams Install and initial openSUSE support
Microsoft Teams is now added as an option to install. Along with that, initial stages of openSUSE support have been added. I don't have a running instance of openSUSE to test with, so the testing will come later. Lastly, rather than just assume that the user would like numlock turned on they will now get asked if they want it turned on. If they say yes, then it will install numlockx and turn it on.